Fix `translation,` and `(translation)`
The first version stored empty dynamic strings by default and used them
to check the unset translations. This made it impossible to use empty
strings as actual translations, which sometimes is useful.

The fixed version stores zeros instead and replaces them with the
address of a dynamic string only when the translation is provided.