Define string constants with Galope's `sconstant`
`2constant` worked because Gforth creates the strings in the common
heap. But `s"` and `s+` had already been redefined with its `stringer`
equivalents in order to reduce as much as possible the usage of the
heap, to fix the memory allocation error that is thrown when many pages
are built at once (e.g. the whole Spanish section).